backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
requests==2.31.0
openai>=1.0.0
beautifulsoup4==4.12.3
python-dotenv==1.0.0
pydantic==2.6.0
httpx>=0.25.0
